requirements.txt
streamlit>=1.22.0
tensorflow>=2.12.0
numpy>=1.22.0
Pillow>=9.0.0
keras
opencv-python-headless>=4.7.0
scikit-image>=0.19.0